+++ src/Crowdin.Api/Reports/GroupTranslationCostGenerateGroupReportRequest.cs
@@ -27,7 +27,7 @@ namespace Crowdin.Api.Reports
         public class RequestSchema
         {
             [JsonProperty("projectIds")]
-            public ICollection<int>? ProjectIds { get; set; }
+            public ICollection<long>? ProjectIds { get; set; }
             
             [JsonProperty("unit")]
             public ReportUnit? Unit { get; set; }
@@ -51,7 +51,7 @@ namespace Crowdin.Api.Reports
             public DateTimeOffset? DateTo { get; set; }
             
             [JsonProperty("userIds")]
-            public ICollection<int>? UserIds { get; set; }
+            public ICollection<long>? UserIds { get; set; }
         }
     }
 }
